Add GitHub Actions Version Updater workflow
This commit adds a new file `.github/workflows/updater.yml` which contains the configuration for a GitHub Actions workflow. The workflow is triggered on a schedule, running every Sunday at midnight. It also has the option to be manually triggered. The job runs on an Ubuntu environment and consists of two steps: checking out the repository using `actions/checkout@v4.1.0`, and running the GitHub Actions Version Updater using `saadmk11/github-actions-version-updater@v0.8.1`. Both steps require an access token with `workflow` scope, which is provided as a secret named `CWB_TOKEN`.